register functionality

+++ Backend/property-management/src/main/kotlin/com/example/propertymanagement/model/Trustee.kt
@@ -6,6 +6,8 @@ import jakarta.persistence.GeneratedValue
 import jakarta.persistence.GenerationType
 import jakarta.persistence.Id
 import jakarta.persistence.Table
+import jakarta.persistence.*
+
 
 @Entity
 @Table(name = "trustee")
@@ -13,7 +15,7 @@ data class Trustee(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "trustee_id")
-    val trustee_id: Int = 0,
+    val trustee_id: Long = 0,
     val name: String,
     val email: String,
     val phone: String,
